検査の陽性率: Rename components/PositiveRateMixedChart.vue to components/index/CardsMonitoring/PositiveRate/Chart.vue
