+++ components/index/CardsMonitoring/PositiveRate/Card.vue
@@ -1,7 +1,7 @@
 <template>
   <v-col cols="12" md="6" class="DataCard PositiveRateCard">
     <client-only>
-      <positive-rate-mixed-chart
+      <chart
         :title="$t('モニタリング項目(4)')"
         :title-id="'positive-rate'"
         :info-titles="[$t('検査の陽性率'), $t('検査人数')]"
@@ -69,7 +69,7 @@
             </li>
           </ul>
         </template>
-      </positive-rate-mixed-chart>
+      </chart>
     </client-only>
   </v-col>
 </template>
@@ -78,7 +78,7 @@
 import dayjs from 'dayjs'
 import duration from 'dayjs/plugin/duration'
 
-import PositiveRateMixedChart from '@/components/PositiveRateMixedChart'
+import Chart from '@/components/index/CardsMonitoring/PositiveRate/Chart.vue'
 import PositiveRate from '@/data/positive_rate.json'
 import {
   getCommaSeparatedNumberToFixedFunction,
@@ -89,7 +89,7 @@ dayjs.extend(duration)
 
 export default {
   components: {
-    PositiveRateMixedChart,
+    Chart,
   },
   data() {
     // 検査実施日別状況
