+++ components/index/CardsMonitoring/PositiveRate/Chart.vue
